* master: (340 commits)
  Only updating premium expire when checking.
  Adding error handling for new icon feed link fetcher.
  Fixing rename title update on ipad.
  Updating current feed's title on rename.
  Adding link to Facebook page.
  Attempting to fetch Tumblr feed icons.
  Auto-enabling open first unread story setting.
  Adding a never_expire flag when upgrading premiums.
  Don't bother recalculating feed's subs, as it will do that on next fetch.
  Changing premium expiration tasks to not get bogged down by re-downloading all premium history beforehand.
  Fixing YouTube videos that weren't displaying in single story mode in Split view. Needs rendering to be handled properly when switching views.
  Fixing list_id
  Motd.
  Fixing YouTube playlist urls.
  Cleaning up fabfile a bit.
  Adding duration and channel info to youtube videos.
  Adding channel_id youtube.com video feeds.
  Allowing fullscreen videos for youtube.
  Upgrading HAProxy to 1.5.12.
  Upgrading readability library.
  ...

+++ apps/search/models.py
@@ -140,14 +140,18 @@ class MUserSearch(mongo.Document):
             feed.index_stories_for_search()
         
     @classmethod
-    def remove_all(cls):
+    def remove_all(cls, drop_index=False):
         user_searches = cls.objects.all()
         logging.info(" ---> ~SN~FRRemoving ~SB%s~SN user searches..." % user_searches.count())
         for user_search in user_searches:
-            user_search.remove()
-            
-        logging.info(" ---> ~FRRemoving stories search index...")
-        SearchStory.drop()
+            try:
+                user_search.remove()
+            except Exception, e:
+                print " ****> Error on search removal: %s" % e
+        
+        if drop_index:
+            logging.info(" ---> ~FRRemoving stories search index...")
+            SearchStory.drop()
         
     def remove(self):
         from apps.rss_feeds.models import Feed
@@ -369,7 +373,6 @@ class SearchFeed:
         }
         cls.ES.indices.put_mapping("%s-type" % cls.name, {
             'properties': mapping,
-            '_source': {'enabled': False},
         }, ["%s-index" % cls.name])
         
     @classmethod
@@ -388,6 +391,7 @@ class SearchFeed:
         
     @classmethod
     def query(cls, text):
+        cls.create_elasticsearch_mapping()
         try:
             cls.ES.default_indices = cls.index_name()
             cls.ES.indices.refresh()
